support voor wachtruimte

+++ src/backend/controllers/classes/teachers/teachersController.ts
@@ -7,7 +7,7 @@ import {
     doesTokenBelongToTeacherInClass,
     getJWToken,
 } from "../../authentication/extraAuthentication.ts";
-import {splitId, teacherLink} from "../../../help/links.ts";
+import {splitId, studentLink, teacherLink} from "../../../help/links.ts";
 import {zTeacherLink} from "../../../help/validation.ts";
 
 export async function getClassTeachers(req: Request, res: Response, next: NextFunction) {
@@ -20,13 +20,19 @@ export async function getClassTeachers(req: Request, res: Response, next: NextFu
     if (!(auth1.success || auth2.success))
         return throwExpressException(403, auth1.errorMessage + " and " + auth2.errorMessage, next);
 
-    //class exist check done by auth
-
     const teachers = await prisma.classTeacher.findMany({
         where: {classes_id: classId.data, accepted: true},
     });
+    const waitingroom = await prisma.classTeacher.findMany({
+        where: {classes_id: classId.data, accepted: false},
+    });
+
     const teacherLinks = teachers.map(teacher => teacherLink(teacher.teachers_id));
-    res.status(200).send({teachers: teacherLinks});
+    const waitingroomlinks = waitingroom.map(teacher => teacherLink(teacher.teachers_id));
+    res.status(200).send({
+        teachers: teacherLinks,
+        waitingroom: waitingroomlinks,
+    });
 }
 
 export async function postClassTeacher(req: Request, res: Response, next: NextFunction) {
@@ -50,30 +56,30 @@ export async function postClassTeacher(req: Request, res: Response, next: NextFu
     res.status(200).send();
 }
 
-// export async function patchClassTeacher(req: Request, res: Response, next: NextFunction) {
-//     const classId = z.coerce.number().safeParse(req.params.classId);
-//     const teacherId = z.coerce.number().safeParse(req.params.teacherId);
-//     if (!classId.success) return throwExpressException(400, "invalid classId", next);
-//     if (!teacherId.success) return throwExpressException(400, "invalid teacherLink", next);
-//
-//     const token = getJWToken(req, next);
-//     const auth1 = await doesTokenBelongToTeacherInClass(classId.data, token);
-//     if (!auth1.success) return throwExpressException(403, auth1.errorMessage, next);
-//
-//     await prisma.classTeacher.update({
-//         where: {
-//             classes_id_teachers_id: {
-//                 classes_id: classId.data,
-//                 teachers_id: teacherId.data
-//             }
-//         },
-//         data: {
-//             accepted: true
-//         }
-//     })
-//
-//     res.status(200).send();
-// }
+export async function patchClassTeacher(req: Request, res: Response, next: NextFunction) {
+    const classId = z.coerce.number().safeParse(req.params.classId);
+    const teacherId = z.coerce.number().safeParse(req.params.teacherId);
+    if (!classId.success) return throwExpressException(400, "invalid classId", next);
+    if (!teacherId.success) return throwExpressException(400, "invalid teacher", next);
+
+    const token = getJWToken(req, next);
+    const auth1 = await doesTokenBelongToTeacherInClass(classId.data, token);
+    if (!auth1.success) return throwExpressException(403, auth1.errorMessage, next);
+
+    await prisma.classTeacher.update({
+        where: {
+            classes_id_teachers_id: {
+                classes_id: classId.data,
+                teachers_id: teacherId.data,
+            }
+        },
+        data: {
+            accepted: true
+        }
+    })
+
+    res.status(200).send();
+}
 
 export async function deleteClassTeacher(req: Request, res: Response, next: NextFunction) {
     const classId = z.coerce.number().safeParse(req.params.classId);
@@ -87,20 +93,23 @@ export async function deleteClassTeacher(req: Request, res: Response, next: Next
 
     //no class or teacher exist checks needed because auth already does this
 
-    await prisma.$transaction([
-        prisma.classTeacher.deleteMany({
-            where: {
+
+    const teacher = await prisma.classTeacher.findUnique({
+        where: {
+            classes_id_teachers_id: {
                 classes_id: classId.data,
-                teachers_id: teacherId.data
+                teachers_id: teacherId.data,
             }
-        }),
-        //verwijder een class als er geen teachers meer voor zijn
-        prisma.class.deleteMany({
-            where: {
-                id: classId.data,
-                classes_teachers: {none: {}}
-            }
-        }),
-    ]);
+        }
+    });
+    if (!teacher) return throwExpressException(404, "teacher not found", next);
+
+    await prisma.classTeacher.deleteMany({
+        where: {
+            classes_id: classId.data,
+            teachers_id: teacherId.data
+        }
+    });
+
     res.status(200).send();
 }
